+++ clients/simple-earn/src/rest-api/types/get-simple-earn-locked-product-list-response-rows-inner-detail.ts
@@ -63,10 +63,10 @@ export interface GetSimpleEarnLockedProductListResponseRowsInnerDetail {
     status?: string;
     /**
      *
-     * @type {string}
+     * @type {number}
      * @memberof GetSimpleEarnLockedProductListResponseRowsInnerDetail
      */
-    subscriptionStartTime?: string;
+    subscriptionStartTime?: number;
     /**
      *
      * @type {string}